Factor out getting meta as YAML from table into meta.py module
